Moving common and character utils to separate files
